+++ plugins/source/k8s/resources/services/core/pods.go
@@ -98,10 +98,10 @@ func Pods() *schema.Table {
 				IgnoreInTests: true,
 			},
 			{
-				Name:        "cluster_name",
-				Description: "The name of the cluster which the object belongs to.",
+				Name:        "zzz_cluster_name",
+				Description: "Deprecated: ClusterName is a legacy field that was always cleared by the system and never used; it will be removed in the future. The name in the database is changed to help clients detect accidental use.",
 				Type:        schema.TypeString,
-				Resolver:    schema.PathResolver("ObjectMeta.ClusterName"),
+				Resolver:    schema.PathResolver("ObjectMeta.ZZZ_DeprecatedClusterName"),
 			},
 			{
 				Name:        "restart_policy",