+++ website/pages/_app.tsx
@@ -1,0 +1,22 @@
+import "../styles.css";
+import "nextra-theme-docs/style.css";
+import "../custom.css";
+import { Prism } from "prism-react-renderer";
+
+(typeof global !== "undefined" ? global : window).Prism = Prism;
+
+require("prismjs/components/prism-powershell");
+require("prismjs/components/prism-hcl");
+require("prismjs/components/prism-ini");
+require("prismjs/components/prism-docker");
+require("prismjs/components/prism-toml");
+
+// Shim requestIdleCallback in Safari
+if (typeof window !== "undefined" && !("requestIdleCallback" in window)) {
+  window.requestIdleCallback = (fn) => setTimeout(fn, 1);
+  window.cancelIdleCallback = (e) => clearTimeout(e);
+}
+
+export default function Nextra({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}